success and cancel notifiations on login, logout, add and delete from cart

+++ CyberPlus/src/app/components/user/product/product.component.ts
@@ -1,5 +1,6 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { CartService } from 'src/app/shared/services/cart.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-product',
@@ -20,8 +21,31 @@ export class ProductComponent {
   }
 
   removeFromCart(id: number) {
-    this.cartService.deleteFromCart(id);
-    this.cart = this.cartService.getCart()
+    Swal.fire({
+      position: 'bottom',
+      icon: 'question',
+      title: 'No se ha podido iniciar sesión',
+      cancelButtonText: 'Cancelar',
+      confirmButtonText: 'Borrar',
+      confirmButtonColor: '#3085d6',
+      showCancelButton: true,
+      cancelButtonColor: '#d33',
+      reverseButtons: true,
+      toast: true,
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.cartService.deleteFromCart(id);
+        this.cart = this.cartService.getCart()
+        Swal.fire({
+          position: 'bottom',
+          title: 'Eliminado!',
+          icon: 'success',
+          toast: true,
+          timer: 1000,
+          showConfirmButton: false
+        })
+      }
+    })
   }
 
 }
